requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
click==8.1.3
cloudpickle==2.2.1
cma==3.3.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.6
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
fasteners==0.18
fonttools==4.38.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
glfw==2.5.6
GPy==1.10.0
GPyOpt==1.2.6
gym==0.26.2
gym-notices==0.0.8
hydra-core==1.3.2
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
moviepy==1.0.3
mujoco-py==2.1.2.14
mypy-extensions==1.0.0
omegaconf==2.3.0
packaging==23.0
paramz==0.9.5
pathspec==0.11.0
pathtools==0.1.2
Pillow==9.4.0
ply==3.11
POT==0.8.2
proglog==0.1.10
protobuf==4.21.12
psutil==5.9.4
pycparser==2.21
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyswarms==1.3.0
python-dateutil==2.8.2
PyYAML==6.0
scikit-learn==1.2.1
scipy==1.10.0
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
termcolor==2.3.0
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
wandb==0.15.2
webencodings==0.5.1
